refactor: restructure `ui` components into individual folders and update imports across app
