+++ apps/frontend/src/components/HeaderSection/HeaderSection.tsx
@@ -1,6 +1,6 @@
-import Box from '@/ui/Box';
-import Button from '@/ui/Button';
-import Typography from '@/ui/Typography';
+import Box from '@/ui/Box/Box.tsx';
+import Button from '@/ui/Button/Button.tsx';
+import Typography from '@/ui/Typography/Typography.tsx';
 import Add from '@mui/icons-material/Add';
 import Delete from '@mui/icons-material/Delete';
