+++ apps/frontend/src/components/TrackForm/TrackForm.tsx
@@ -1,14 +1,14 @@
 import { useEffect } from 'react';
 import { useForm, Controller } from 'react-hook-form';
-import Dialog from '@/ui/Dialog';
-import DialogTitle from '@/ui/DialogTitle';
-import DialogContent from '@/ui/DialogContent';
-import TextField from '@/ui/TextField';
-import Autocomplete from '@/ui/Autocomplete';
-import Chip from '@/ui/Chip';
-import Button from '@/ui/Button';
-import DialogActions from '@/ui/DialogActions';
-import IconButton from '@/ui/IconButton';
+import Dialog from '@/ui/Dialog/Dialog.tsx';
+import DialogTitle from '@/ui/DialogTitle/DialogTitle.tsx';
+import DialogContent from '@/ui/DialogContent/DialogContent.tsx';
+import TextField from '@/ui/TextField/TextField.tsx';
+import Autocomplete from '@/ui/Autocomplete/Autocomplete.tsx';
+import Chip from '@/ui/Chip/Chip.tsx';
+import Button from '@/ui/Button/Button.tsx';
+import DialogActions from '@/ui/DialogActions/DialogActions.tsx';
+import IconButton from '@/ui/IconButton/IconButton.tsx';
 import { ITrackForm, defaultGenres } from './Interface';
 import { TTrackFormData } from '@/types/types.ts';
 import Close from '@mui/icons-material/Close';
