+++ apps/frontend/src/components/TrackItem/components/TrackInfo/TrackInfo.tsx
@@ -1,15 +1,25 @@
-import Box from '@/ui/Box';
-import Typography from '@/ui/Typography';
-import Chip from '@/ui/Chip';
+import Box from '@/ui/Box/Box.tsx';
+import Typography from '@/ui/Typography/Typography.tsx';
+import Chip from '@/ui/Chip/Chip.tsx';
 import { ITrack } from '@/types/types.ts';
 import React, { Suspense } from 'react';
 import LoadingIndicator from '@/components/LoadingIndicator/LoadingIndicator.tsx';
-const CoverImage = React.lazy(() => import('@/ui/Avatar'));
+const CoverImage = React.lazy(() => import('@/ui/Avatar/Avatar.tsx'));
 
-const TrackInfo = ({ track, loading }: { track: ITrack; loading?: 'eager' | 'lazy' }) => (
+const TrackInfo = ({
+  track,
+  loading,
+}: {
+  track: ITrack;
+  loading?: 'eager' | 'lazy';
+}) => (
   <Box display="flex" alignItems="center" gap={2}>
     <Suspense fallback={<LoadingIndicator size={20} message={''} />}>
-      <CoverImage src={track.coverImage} sx={{ width: 50, height: 50 }} {...(loading ? { imgProps: { loading } } : {})} />
+      <CoverImage
+        src={track.coverImage}
+        sx={{ width: 50, height: 50 }}
+        {...(loading ? { imgProps: { loading } } : {})}
+      />
     </Suspense>
     <Box>
       <Typography
