+++ apps/frontend/src/components/TrackList/TrackList.tsx
@@ -1,4 +1,4 @@
-import Grid from '@/ui/Grid';
+import Grid from '@/ui/Grid/Grid.tsx';
 import TrackItem from '../TrackItem/TrackItem.tsx';
 import TrackListSkeleton from './TrackListSkeleton';
 import { FC } from 'react';
@@ -34,29 +34,29 @@ const TrackList: FC<TrackListProps> = ({
   if (isLoading) return <TrackListSkeleton />;
 
   return (
-  <Grid container spacing={3} data-testid="tracks-list" direction="column">
+    <Grid container spacing={3} data-testid="tracks-list" direction="column">
       {tracksData?.tracks.map((track, idx) => (
         <Grid key={track.id} data-testid={`track-item-${track.id}`}>
-        <TrackItem
-          track={track}
-          onEdit={() => onEditTrack(track.id)}
-          onDelete={() => onDeleteTrack(track.id)}
-          isSelectMode={isSelectMode}
-          isSelected={selectedTracks.includes(track.id)}
-          onSelect={() => onSelectTrack(track.id)}
+          <TrackItem
+            track={track}
+            onEdit={() => onEditTrack(track.id)}
+            onDelete={() => onDeleteTrack(track.id)}
+            isSelectMode={isSelectMode}
+            isSelected={selectedTracks.includes(track.id)}
+            onSelect={() => onSelectTrack(track.id)}
             isLoading={false}
-          data-testid={{
-            title: `track-item-${track.id}-title`,
-            artist: `track-item-${track.id}-artist`,
-            edit: `edit-track-${track.id}`,
-            delete: `delete-track-${track.id}`,
-          }}
+            data-testid={{
+              title: `track-item-${track.id}-title`,
+              artist: `track-item-${track.id}-artist`,
+              edit: `edit-track-${track.id}`,
+              delete: `delete-track-${track.id}`,
+            }}
             loading={idx === 0 ? 'eager' : undefined}
-        />
-      </Grid>
-    ))}
-  </Grid>
-);
+          />
+        </Grid>
+      ))}
+    </Grid>
+  );
 };
 
 export default TrackList;
