+++ apps/frontend/src/pages/TracksPage/TracksPage.tsx
@@ -1,5 +1,5 @@
-import Container from '@/ui/Container';
-import Paper from '@/ui/Paper';
+import Container from '@/ui/Container/Container.tsx';
+import Paper from '@/ui/Paper/Paper.tsx';
 import { useTrackPageState } from '@/hooks';
 import HeaderSection from '@/components/HeaderSection/HeaderSection.tsx';
 import FiltersAndBulk from './components/FiltersAndBulk/FiltersAndBulk.tsx';
@@ -7,7 +7,9 @@ import Tracks from './components/Tracks/Tracks.tsx';
 import React, { Suspense } from 'react';
 import LoadingIndicator from '@/components/LoadingIndicator/LoadingIndicator.tsx';
 
-const TrackModals = React.lazy(() => import('./components/TrackModals/TrackModals'));
+const TrackModals = React.lazy(
+  () => import('./components/TrackModals/TrackModals')
+);
 
 const TrackPage = () => {
   const state = useTrackPageState();
@@ -45,7 +47,7 @@ const TrackPage = () => {
         <Tracks state={state} />
       </Paper>
       <Suspense fallback={<LoadingIndicator message="Loading modals..." />}>
-      <TrackModals state={state} />
+        <TrackModals state={state} />
       </Suspense>
     </Container>
   );
