+++ apps/frontend/src/ui/Avatar/Avatar.tsx
