+++ apps/frontend/src/ui/DialogContent/DialogContent.tsx
