+++ apps/frontend/src/ui/index.ts
@@ -1,18 +1,18 @@
-export { default as Button } from './Button';
-export { default as Box } from './Box';
-export { default as Dialog } from './Dialog';
-export { default as DialogTitle } from './DialogTitle';
-export { default as DialogContent } from './DialogContent';
-export { default as DialogActions } from './DialogActions';
-export { default as IconButton } from './IconButton';
-export { default as Checkbox } from './Checkbox';
-export { default as FormControlLabel } from './FormControlLabel';
-export { default as TextField } from './TextField';
-export { default as Autocomplete } from './Autocomplete';
-export { default as Avatar } from './Avatar';
-export { default as Chip } from './Chip';
-export { default as Typography } from './Typography';
-export { default as CircularProgress } from './CircularProgress';
-export { default as Container } from './Container';
-export { default as Paper } from './Paper';
-export { default as Grid } from './Grid';
+export { default as Button } from './Button/Button.tsx';
+export { default as Box } from './Box/Box.tsx';
+export { default as Dialog } from './Dialog/Dialog.tsx';
+export { default as DialogTitle } from './DialogTitle/DialogTitle.tsx';
+export { default as DialogContent } from './DialogContent/DialogContent.tsx';
+export { default as DialogActions } from './DialogActions/DialogActions.tsx';
+export { default as IconButton } from './IconButton/IconButton.tsx';
+export { default as Checkbox } from './Checkbox/Checkbox.tsx';
+export { default as FormControlLabel } from './FormControlLabel/FormControlLabel.tsx';
+export { default as TextField } from './TextField/TextField.tsx';
+export { default as Autocomplete } from './Autocomplete/Autocomplete.tsx';
+export { default as Avatar } from './Avatar/Avatar.tsx';
+export { default as Chip } from './Chip/Chip.tsx';
+export { default as Typography } from './Typography/Typography.tsx';
+export { default as CircularProgress } from './CircularProgress/CircularProgress.tsx';
+export { default as Container } from './Container/Container.tsx';
+export { default as Paper } from './Paper/Paper.tsx';
+export { default as Grid } from './Grid/Grid.tsx';
